Begin first day of AOC
Spent the first few hours learning about Go data types, specifically the differences between hash maps and slices. It’s interesting to learn that slices often (depending on escape analysis) live on the stack, making access more efficient since their elements are stored contiguously in memory. In contrast, hash maps typically reside in the heap, and accessing their values involves pointer dereferencing and bucket management, which adds overhead. For these reasons, slices are generally a more efficient data structure, especially when working with two arrays of integers.

That said, this difference might be trivial when dealing with a dataset like input.txt containing up to 2000 numbers, as this is relatively small. I also found it interesting that if the dataset grows too large (e.g. several kbs), slices may be allocated on the heap instead of the stack to accommodate their size.